+++ src/Main.java
@@ -1,0 +1,8 @@
+public class Main {
+
+    static SimpleBot simpleBot = new SimpleBot();
+
+    public static void main(String[] args) {
+        simpleBot.run();
+    }
+}
